move updateTeamMember to own file
